server/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pillow
pydantic